requirements.txt
africastalking==2.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
mccabe==0.7.0
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.5
responses==0.25.8
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
